refactor: break colonizethis_logic ↔ colonizethis_ai cycle (ref #1531)
- Move AIConfig, AISeedBundle, EconomyPlan, CargoPreference, StrategicOrderResult,
  and AssignedRecipe into colonizethis_models.
- Extract OrderSuggestionAPI to order_suggestion_api library; remove logic pub dep on ai.
- Move assignHiddenAgendasForGame and kHiddenAgendaIds to logic; keep agenda modifiers in ai.
- Move generateOrdersForPlayerFullAI / generateOrdersForGameFullAI / FullAIResult to ai.
- Remove colonizethis_ai di re-export; tests import logic di directly.
- Re-export shared types from colonizethis_ai barrel for compatibility.

Made-with: Cursor
