+++ app/lib/features/game/flame/game_screen.dart
@@ -1,5 +1,6 @@
 export 'game_screen_shared.dart';
 
+import 'package:colonizethis_ai/colonizethis_ai.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
