+++ app/test/game_screen_next_turn_ai_test.dart
@@ -1,4 +1,5 @@
 import 'package:colonizethis_app/features/game/flame/game_screen.dart';
+import 'package:colonizethis_ai/colonizethis_ai.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
