+++ packages/colonizethis_ai/lib/colonizethis_ai.dart
@@ -1,14 +1,21 @@
 // AI behavior, planning, personalities. SPEC/ai/ai-architecture.md, SPEC/program/ai-systems-impl.md.
 
 export 'package:colonizethis_data/colonizethis_data.dart' show kPortraitMoodValues;
-export 'src/ai_config.dart';
+export 'package:colonizethis_models/colonizethis_models.dart'
+    show
+        AIConfig,
+        AISeedBundle,
+        AssignedRecipe,
+        CargoPreference,
+        EconomyPlan,
+        StrategicOrderResult;
 export 'src/dossier.dart';
 export 'src/domain_planners.dart';
 export 'src/economy_planner.dart';
+export 'src/full_ai_planner.dart';
 export 'src/goal_manager.dart';
 export 'src/hidden_agenda.dart';
 export 'src/mood_state_machine.dart';
 export 'src/perception.dart';
-export 'src/seed_bundle.dart';
 export 'src/strategic_ai.dart';
 export 'src/tactical_ai.dart';
