--- packages/colonizethis_ai/lib/di.dart
@@ -1,3 +1,0 @@
-// Re-exports logic-layer Riverpod providers for AI tests and composition. See SPEC/program/dependency-injection.md.
-export 'package:colonizethis_logic/di.dart'
-    show gameEventBusProvider, orderSuggestionApiProvider;
