+++ packages/colonizethis_ai/test/ai_di_export_test.dart
@@ -1,9 +1,9 @@
-import 'package:colonizethis_ai/di.dart';
+import 'package:colonizethis_logic/di.dart';
 import 'package:colonizethis_test/test.dart';
 import 'package:riverpod/riverpod.dart';
 
 void main() {
-  test('di re-exports orderSuggestionApiProvider', () {
+  test('logic di exposes orderSuggestionApiProvider for AI composition', () {
     final container = ProviderContainer();
     addTearDown(container.dispose);
     expect(container.read(orderSuggestionApiProvider), isNotNull);
