+++ packages/colonizethis_logic/lib/order_suggestion_api.dart
@@ -1,0 +1,4 @@
+/// Narrow export: [OrderSuggestionAPI] only (for colonizethis_ai). SPEC/program/dependency-injection.md.
+library;
+
+export 'src/orders/order_suggestion_api.dart';
