+++ packages/colonizethis_logic/lib/src/di/logic_providers.dart
@@ -1,7 +1,7 @@
 import 'package:riverpod/riverpod.dart';
 
 import '../event_bus/game_event_bus.dart';
-import '../orders/order_suggestion.dart';
+import '../orders/order_suggestion_api.dart';
 import '../orders/order_suggestion_api_impl.dart';
 
 /// Default [OrderSuggestionAPI] for AI and tooling. Override in tests via [ProviderContainer].
