+++ packages/colonizethis_logic/pubspec.yaml
@@ -12,12 +12,12 @@ dependencies:
   logger: ^2.0.2
   riverpod: ^3.2.1
   colonizethis_logger:
-  colonizethis_ai:
   colonizethis_data:
   colonizethis_map:
   colonizethis_models:
 
 dev_dependencies:
+  colonizethis_ai:
   colonizethis_test:
   coverage: ^1.15.0
   test: ^1.24.0
