+++ packages/colonizethis_logic/test/logic_providers_test.dart
@@ -1,4 +1,5 @@
 import 'package:colonizethis_test/test.dart';
+import 'package:colonizethis_ai/colonizethis_ai.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_logic/di.dart';
