+++ packages/colonizethis_models/lib/colonizethis_models.dart
@@ -30,6 +30,11 @@ export 'src/turn_time_mapping.dart';
 export 'src/unit.dart';
 export 'src/world_state.dart';
 export 'src/ai_events.dart';
+export 'src/assigned_recipe.dart';
+export 'src/ai_config.dart';
+export 'src/ai_seed_bundle.dart';
+export 'src/economy_plan.dart';
+export 'src/strategic_order_result.dart';
 export 'src/app_event_bus.dart';
 export 'src/app_events.dart';
 export 'src/dossier_evidence.dart';
