+++ packages/colonizethis_models/lib/src/ai_config.dart
