+++ packages/colonizethis_models/lib/src/ai_seed_bundle.dart
